refactoring API call
