+++ src/components/Form.js
@@ -1,11 +1,11 @@
-import { useState } from "react";
+import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import axios from "axios";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSearch } from "@fortawesome/free-solid-svg-icons";
 
-const Form = ({ name, setSearchedFilms, setUserInput, userInput }) => {
-  // const [userInput, setUserInput] = useState("");
+const Form = ({ name, setSearchedFilms }) => {
+  const [userInput, setUserInput] = useState("");
 
   // const [searchedFilms, setSearchedFilms] = useState([]);
 
@@ -55,9 +55,7 @@ const Form = ({ name, setSearchedFilms, setUserInput, userInput }) => {
   const handleSubmit = (event) => {
     fetchSearchMovies(userInput);
     event.preventDefault();
-    setTimeout(() => {
-      setUserInput("");
-    }, 10000); // to make the input text dissapear 
+    setUserInput("");
   };
 
   const handleUserInput = (event) => {
@@ -66,11 +64,7 @@ const Form = ({ name, setSearchedFilms, setUserInput, userInput }) => {
   };
 
   return (
-    <div className="formContainer">
-      <div className="formHeader">
-        <h2>Welcome to The Film Factory!</h2>
-        <h3> Search your favourite English language film, get results from around the world!</h3>
-      </div>
+    <section className="formContainer">
       <form onSubmit={handleSubmit}>
         <legend>What are you looking for?</legend>
         <label htmlFor="search"></label>
@@ -85,12 +79,14 @@ const Form = ({ name, setSearchedFilms, setUserInput, userInput }) => {
             placeholder=" What do you have in mind? "
           />
         </div>
-
-        <button className="searchButton">Search</button>
       </form>
-      
-    </div>
+      <div className="buttonContainer">
+        <button className="searchButton">Search</button>
+      </div>
+    </section>
+    
   );
 };
 
 export default Form;
+
